+++ CHANGELOG.md
@@ -3,6 +3,9 @@ This document describes changes between versions of rbxmk. The `imperative`
 branch is the latest unreleased version.
 
 ## imperative
+**Highlights:**
+- Add [os.clean](https://github.com/Anaminus/rbxmk/blob/imperative/doc/libraries.md#user-content-osclean) function to clean file paths.
+
 **Fixes:**
 - Fix missing properties from sym.Properties when instance has a descriptor.
 - Fix unexpected conversion failures when getting certain property types while